backporting msm/kgsl updates from codeaurora msm-3.4 - part 7 ___________
msm: kgsl: Declare some mmu functions as static inline

Declare those functions that simply call the mmu-specific function
pointer as static inline functions in the mmu header file. Also,
change the return type of the mmu stop function to void.

Change-Id: I4efd7901858c768eec16886c8410d362b918489b
Signed-off-by: Shubhraprakash Das <sadas@codeaurora.org>
________________

msm: kgsl: Increase the virtual GPU range for the IOMMU

Instead of a fixed 256MB virtual range for both the GPUMMU and IOMMU, make
the virtual range a property of the MMU engine and set the IOMMU range to
2GB.  Technically we could go all the way up to 4G, but even 2G is far out
of the realm of possiblity in the current generation, and we wanted to
reserve some of the space for future enhancements.

Change-Id: Ic0dedbad2987beb162b6a1878dd65ffae8a78522
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>